Create abstract_class.py
Abstract class is a class where you initialize methods which are in higher abstraction and use these methods to perform operations in the child class. 

Use @staticmethod decorator to let the tool know that it's the abstract method and this makes the class an abstract class (a class which does not need any object).